+++ R/valid.R
@@ -1,0 +1,7 @@
+valid__name_types <- function() {
+  c(
+    "internal",
+    "external",
+    "external_clean"
+  )
+}