+++ NEWS.md
@@ -1,3 +1,10 @@
+# namap 0.0.0.9001
+
+- Added function `get_name()`
+- Added function `get_names()`
+
+--------------------------------------------------------------------------------
+
 # namap 0.0.0.9000
 
 - Original commit